requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.1.7
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-webpack-loader==0.6.0
gunicorn==19.9.0
html5lib==0.999
httplib2==0.9.1
idna==2.8
Jinja2==2.8
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pusher==2.1.4
pyasn1==0.4.6
pycodestyle==2.5.0
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dotenv==0.10.1
pytz==2018.9
requests==2.22.0
six==1.12.0
urllib3==1.25.3
